Site works in Postman and can get list of DB at http://localhost:8080/persons
Added Lombrok to minimize poerson.java.
