+++ src/main/java/projekt/sysjm3/rest/RestApplication.java
@@ -9,6 +9,8 @@ public class RestApplication {
 	public static void main(String[] args) {
 		SpringApplication.run(RestApplication.class, args);
 		System.out.println("TEST");
+		
+		
 	}
 
 }
