+++ src/main/java/projekt/sysjm3/rest/Service/PersonService.java
@@ -6,30 +6,31 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+
 import projekt.sysjm3.rest.Entity.Person;
 import projekt.sysjm3.rest.Repository.PersonRepository;
 
 @Service
 public class PersonService {
-	
-	@Autowired
-	PersonRepository personRepository;
 
-	public List<Person> getAllPersons() {
-		List<Person> persons = new ArrayList<Person>();
-		personRepository.findAll().forEach(person -> persons.add(person));
-		return persons;
-	}
+    @Autowired
+    PersonRepository personRepository;
 
-	public Person getPersonById(int id) {
-		return personRepository.findById(id).get();
-	}
+    public List<Person> getAllPersons() {
+        List<Person> persons = new ArrayList<Person>();
+        personRepository.findAll().forEach(person -> persons.add(person));
+        return persons;
+    }
 
-	public void saveOrUpdate(Person person) {
-		personRepository.save(person);
-	}
+    public Person getPersonById(int id) {
+        return personRepository.findById(id).get();
+    }
 
-	public void delete(int id) {
-		personRepository.deleteById(id);
-	}	
+    public void saveOrUpdate(Person person) {
+        personRepository.save(person);
+    }
+
+    public void delete(int id) {
+        personRepository.deleteById(id);
+    }
 }
