+++ src/main/resources/data.sql
@@ -1,0 +1,3 @@
+insert into person values (1,45,'Manchester','England','Liam','Gallagher',905015);
+insert into person values (2,46,'Manchester','England','Noel','Gallagher',910608);
+insert into person values (3,47,'Manchester','England','Ian','Brown',920512);
