+++ src/main/resources/templates/index.html
@@ -1,62 +1,12 @@
 <!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
+<html xmlns:th="http://www.thymeleaf.org">
 <head>
 
-<meta charset="utf-8" />
-<meta name="viewport"
-	content="width=device-width, initial-scale=1, shrink-to-fit=no">
 
-
-<link rel="stylesheet"
-	href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css"
-	integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm"
-	crossorigin="anonymous">
 <title>Insert title here</title>
 </head>
 <body>
 
 
-
-	<div class="container">
-		<div th:switch="${persons}" class="container">
-			<table class="table">
-<thead>
-<tr>
-<th scope="col">Lastname</th>
-<th scope="col">Firstname</th>
-<th scope="col">Age</th>
-<th scope="col">SSN</th>
-<th scope="col">City</th>
-<th scope="col">Country</th>
-<!-- <th scope="col">Gender</th> -->
-</tr>
-</thead>
-<tbody>
-<tr th:each="person : ${persons}">
-<td th:text="${person.lastname}"></td>
-<td th:text="${person.firstname}"></td>
-<td th:text="${person.age}"></td>
-<td th:text="${person.ssn}"></td>
-<td th:text="${person.city}"></td>
-<td th:text="${person.country}"></td>
-<td><a th:href="@{/edit/{id}(id=${person.id})}" class="btn btn-info">Edit</a></td>
-<td><a th:href="@{/delete/{id}(id=${person.id})}" class="btn btn-info">Edit</a></td>
-
-</tbody>
-</table>
-
-</div>
-	</div>
-	<script src="https://code.jquery.com/jquery-3.2.1.slim.min.js"
-		integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN"
-		crossorigin="anonymous"></script>
-	<script
-		src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js"
-		integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q"
-		crossorigin="anonymous"></script>
-	<script
-		src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js"
-		integrity="sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl"
-		crossorigin="anonymous"></script>
 </body>
 </html>
